Install gassoc and consfile to SYSCONFDIR.
Since these files are customizable by the sysadmin, they are better
installed to sysconfdir.

Reorganize a bit Makefile.am, but the only real change is the location
of the files.

Record this change in the Changelog. This is the real 0.9.3 release. Lucky
me had not tagged anything in svn or uploaded a tarball yet.


git-svn-id: http://grun.googlecode.com/svn/trunk@58 4c072952-ec3e-0410-8942-d3a44bbd62b3